server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Naim <skylike2904@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
mongoengine = "^0.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
